restructured repo, models can now be easier switched around
