+++ tarkibi/__init__.py
@@ -1,2 +1,2 @@
-from .dataset import Tarkibi
+from .tarkibi import Tarkibi
